+++ next-docs/pages/components/button.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 
-import Primary from '../../public/examples/button/Primary';
-import Secondary from '../../public/examples/button/Secondary';
-import Tertiary from '../../public/examples/button/Tertiary';
-import Default from '../../public/examples/button/Default';
-import Sizes from '../../public/examples/button/Sizes';
+import Preview from '../../components/codePreview/Preview';
 import Animations from '../../public/examples/button/Animations';
+import Default from '../../public/examples/button/Default';
+import FullWidth from '../../public/examples/button/FullWidth';
 import Icons from '../../public/examples/button/Icons';
 import IsIcon from '../../public/examples/button/IsIcon';
-import FullWidth from '../../public/examples/button/FullWidth';
+import Primary from '../../public/examples/button/Primary';
+import Secondary from '../../public/examples/button/Secondary';
+import Sizes from '../../public/examples/button/Sizes';
+import Tertiary from '../../public/examples/button/Tertiary';
 
-import Preview from '../../components/codePreview/Preview';
 import useExamples from '../../utils/useExamples';
 
 export default function PageButton() {