+++ next-docs/public/examples/buttonNew/Disabled.tsx
@@ -1,0 +1,9 @@
+import { Button } from '@heathmont/moon-core';
+
+const Example = () => (
+  <div className="flex justify-around w-full items-center">
+    <Button disabled>Disabled</Button>
+  </div>
+);
+
+export default Example;