+++ next-docs/public/examples/select/multiReactSelect/OptionsContent.tsx
@@ -24,11 +24,11 @@ const Example = () => (
     <div className="w-80 mr-4 ml-4 mb-8">
       <Select
         options={options}
-        label="Single Large"
+        label="Large"
         placeholderSlot="Choose an option"
         isMulti
-        menuHeader={<Header />}
-        menuFooter={
+        headerSlot={<Header />}
+        footerSlot={
           <Footer
             primButton={
               <Button fullWidth variant="primary" size="xsmall">
@@ -54,12 +54,12 @@ const Example = () => (
     <div className="w-80 mr-4 ml-4 mb-8">
       <Select
         options={options}
-        label="Single xLarge"
+        label="xLarge"
         size="xLarge"
         placeholderSlot="Choose an option"
         isMulti
-        menuHeader={<Header />}
-        menuFooter={
+        headerSlot={<Header />}
+        footerSlot={
           <Footer
             primButton={
               <Button fullWidth variant="primary" size="xsmall">