+++ packages/core/src/button/private/buttonStyles/buttonCommonStyles.ts
@@ -1,0 +1,41 @@
+import { useTheme } from '@heathmont/moon-themes';
+import { rem } from '@heathmont/moon-utils';
+import type { ButtonProps } from '../../Button';
+
+const buttonCommonStyles = ({ iconSize, disabled }: ButtonProps) => {
+  const { newTokens } = useTheme();
+  return [
+    {
+      position: 'relative',
+      zIndex: 0,
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      fontWeight: newTokens.font.weight.semibold,
+      borderRadius: newTokens.interactiveBorderRadius,
+      textDecoration: 'none',
+      fontFamily: 'inherit',
+      overflow: 'hidden',
+      userSelect: 'none',
+      outline: 'none',
+      transition: `all ${newTokens.transition.default}`,
+      '&:active': {
+        transform: newTokens.transform,
+      },
+      '&:focus-visible': {
+        boxShadow: newTokens.focus,
+      },
+      'img, svg': {
+        display: 'block',
+        width: rem(iconSize),
+        height: rem(iconSize),
+      },
+    },
+    disabled && {
+      opacity: newTokens.opacity,
+      cursor: 'not-allowed',
+    },
+  ];
+};
+
+export default buttonCommonStyles;