+++ packages/select/src/private/MultiValueContainer.tsx
@@ -1,19 +1,18 @@
-// @ts-nocheck
 import React from 'react';
 import { Label } from '@heathmont/moon-core';
 
 //TODO: check Label size according to new size system
-const MultiValueContainer = (props) => {
-  const size = props.selectProps['data-customProps'].size;
+const MultiValueContainer = ({ ...rest }) => {
+  const size = rest.selectProps['data-customProps'].size;
 
   return (
     <Label
       color="gohan.100"
       backgroundColor="piccolo.100"
-      iconRight={props.children[1]}
+      iconRight={rest.children[1]}
       size={size === 'xLarge' ? 'small' : 'medium'}
     >
-      {props.data.label}
+      {rest.data.label}
     </Label>
   );
 };