+++ next-docs/CHANGELOG.md
@@ -1,5 +1,91 @@
 # Change Log
 
+## 7.22.2
+
+### Patch Changes
+
+- Fix assets publish
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@7.22.2
+  - @heathmont/moon-assets@7.22.2
+  - @heathmont/moon-charts@7.22.2
+  - @heathmont/moon-components@7.22.2
+  - @heathmont/moon-core@7.22.2
+  - @heathmont/moon-draggabletable@7.22.2
+  - @heathmont/moon-icons@7.22.2
+  - @heathmont/moon-pagination@7.22.2
+  - @heathmont/moon-popover@7.22.2
+  - @heathmont/moon-select@7.22.2
+  - @heathmont/moon-table@7.22.2
+  - @heathmont/moon-themes@7.22.2
+  - @heathmont/moon-utils@7.22.2
+
+## 7.22.1
+
+### Patch Changes
+
+- New tokens
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@7.22.1
+  - @heathmont/moon-assets@7.22.1
+  - @heathmont/moon-charts@7.22.1
+  - @heathmont/moon-components@7.22.1
+  - @heathmont/moon-core@7.22.1
+  - @heathmont/moon-draggabletable@7.22.1
+  - @heathmont/moon-icons@7.22.1
+  - @heathmont/moon-pagination@7.22.1
+  - @heathmont/moon-popover@7.22.1
+  - @heathmont/moon-select@7.22.1
+  - @heathmont/moon-table@7.22.1
+  - @heathmont/moon-themes@7.22.1
+  - @heathmont/moon-utils@7.22.1
+
+## 7.22.0
+
+### Minor Changes
+
+- add logos
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@7.22.0
+  - @heathmont/moon-core@7.22.0
+  - @heathmont/moon-themes@7.22.0
+  - @heathmont/moon-accordion@7.22.0
+  - @heathmont/moon-charts@7.22.0
+  - @heathmont/moon-components@7.22.0
+  - @heathmont/moon-draggabletable@7.22.0
+  - @heathmont/moon-icons@7.22.0
+  - @heathmont/moon-pagination@7.22.0
+  - @heathmont/moon-popover@7.22.0
+  - @heathmont/moon-select@7.22.0
+  - @heathmont/moon-table@7.22.0
+  - @heathmont/moon-utils@7.22.0
+
+## 7.21.0
+
+### Minor Changes
+
+- New nav & select fixes
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@7.21.0
+  - @heathmont/moon-pagination@7.21.0
+  - @heathmont/moon-popover@7.21.0
+  - @heathmont/moon-select@7.21.0
+  - @heathmont/moon-table@7.21.0
+  - @heathmont/moon-accordion@7.21.0
+  - @heathmont/moon-assets@7.21.0
+  - @heathmont/moon-charts@7.21.0
+  - @heathmont/moon-components@7.21.0
+  - @heathmont/moon-draggabletable@7.21.0
+  - @heathmont/moon-icons@7.21.0
+  - @heathmont/moon-themes@7.21.0
+  - @heathmont/moon-utils@7.21.0
+
 ## 7.20.0
 
 ### Minor Changes